feat(product page): add success state UI in AddToCart
