+++ src/app/sn/products/[productKey]/ProductPageContainer/index.tsx
@@ -1,55 +1,34 @@
 'use client';
 
 import { notFound, useParams } from 'next/navigation';
-import { useShallow } from 'zustand/shallow';
 import { useQuery } from '@tanstack/react-query';
 
 import style from './style.module.scss';
 
-import { useClientStore } from '@/providers/client-store-provider';
-
 import { productOptions } from '@/queries/options';
-import { useCreateCartItemMutation } from '@/queries/hooks/carts/useCreateCartItemMutation';
-
-import Button from '@/components/Button';
 
 import { ProductPageParams } from '../page';
 import ProductInformation from './ProductInformation';
 import ProductRecommend from './ProductRecommend';
 import OtherProducts from './OtherProducts';
+import AddToCart from './AddToCart';
 
 function ProductPageContainer() {
   const params = useParams<ProductPageParams>();
   const { data: product, isError } = useQuery(
     productOptions({ key: params.productKey }),
   );
-  const { mutate } = useCreateCartItemMutation();
-
-  const [addProduct, isLoggedIn] = useClientStore(
-    useShallow((state) => [state.addProductToCart, state.user.isLoggedIn]),
-  );
 
   if (isError) throw new Error('Failed to load product');
   if (!product) return notFound();
 
-  const onAddProduct = () => {
-    const ingredientKey = product.ingredient.key;
-    const productKey = product.key;
-    if (isLoggedIn) {
-      mutate({
-        ingredientKey,
-        productKey,
-      });
-    } else {
-      addProduct(ingredientKey, productKey);
-    }
-  };
   return (
     <div className={style['container']}>
       <div className={style['header']}>
-        <Button onClick={onAddProduct} variant='secondary'>
-          Add to cart
-        </Button>
+        <AddToCart
+          ingredientKey={product.ingredient.key}
+          productKey={product.key}
+        />
       </div>
       <ProductInformation product={product} />
       <OtherProducts />
